+++ packages/leancode_lint/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add new custom lints:
   - [`avoid_catch_error`](https://github.com/leancodepl/flutter_corelibrary/tree/master/packages/leancode_lint#avoid_catch_error)
   - [`missing_equatable_props`](https://github.com/leancodepl/flutter_corelibrary/tree/master/packages/leancode_lint#missing_equatable_props)
+- Replace usages of members deprecated in the analyzer.
 
 # 23.0.0